#660 - store - split claims and compare modules
