+++ src/store/modules/claims/index.js
@@ -1,0 +1,6 @@
+import actions from './actions'
+
+export default {
+  namespaced: true,
+  actions
+}
