+++ src/store/modules/compare/getters.js
@@ -1,0 +1,10 @@
+export default {
+  check (state) {
+    return {
+      isOnCompare: (product) => {
+        let item = state.itemsCompare.find(p => p.sku === product.sku)
+        return (item !== undefined)
+      }
+    }
+  }
+}
