+++ src/store/modules/compare/index.js
@@ -1,0 +1,14 @@
+import actions from './actions'
+import getters from './getters'
+import mutations from './mutations'
+
+export default {
+  namespaced: true,
+  state: {
+    compare: false,
+    itemsCompare: []
+  },
+  getters,
+  actions,
+  mutations
+}
